Move functions out of NetworksDropdown to utils
